+++ samples/document-snippets/instance.js
@@ -1,17 +1,16 @@
-/**
- * Copyright 2018, Google, Inc.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2018 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 const snippets = {
   createInstance: (instanceId, clusterId) => {
@@ -203,6 +202,67 @@ const snippets = {
         // Handle the error.
       });
     // [END bigtable_get_clusters]
+  },
+
+  getIamPolicy: instanceId => {
+    // [START bigtable_get_instance_Iam_policy]
+    const Bigtable = require('@google-cloud/bigtable');
+    const bigtable = new Bigtable();
+    const instance = bigtable.instance(instanceId);
+
+    instance
+      .getIamPolicy()
+      .then(result => {
+        const policy = result[0];
+      })
+      .catch(err => {
+        // Handle the error.
+      });
+    // [END bigtable_get_instance_Iam_policy]
+  },
+
+  setIamPolicy: instanceId => {
+    // [START bigtable_set_instance_Iam_policy]
+    const Bigtable = require('@google-cloud/bigtable');
+    const bigtable = new Bigtable();
+    const instance = bigtable.instance(instanceId);
+
+    const policy = {
+      bindings: [
+        {
+          role: 'roles/bigtable.viewer',
+          members: ['user:mike@example.com', 'group:admins@example.com'],
+        },
+      ],
+    };
+
+    instance
+      .setIamPolicy(policy)
+      .then(result => {
+        const setPolicy = result[0];
+      })
+      .catch(err => {
+        // Handle the error
+      });
+    // [END bigtable_set_instance_Iam_policy]
+  },
+
+  testIamPermissions: instanceId => {
+    // [START bigtable_test_instance_Iam_permissions]
+    const Bigtable = require('@google-cloud/bigtable');
+    const bigtable = new Bigtable();
+    const instance = bigtable.instance(instanceId);
+
+    const permissions = ['bigtable.tables.get', 'bigtable.tables.readRows'];
+    instance
+      .testIamPermissions(permissions)
+      .then(result => {
+        const grantedPermissions = result[0];
+      })
+      .catch(err => {
+        // Handle the error
+      });
+    // [END bigtable_test_instance_Iam_permissions]
   },
 
   getAppProfiles: instanceId => {